Move pipeline injection to its own package
